requirements.txt
numpy>=1.21.0
pandas>=1.3.0
hdbscan>=0.8.27
supabase>=1.0.0
python-dotenv>=0.19.0 